+++ src/Util/createLoggingDefinition.ts
@@ -1,0 +1,15 @@
+import { Definition, DefinitionMembers, LoggingDefinition } from "../definitions";
+
+/** @internal */
+function createLoggingDefinition(definition: Definition | DefinitionMembers) {
+	const loggingDefinition: LoggingDefinition = {
+		Id: (definition as DefinitionMembers).Id,
+		Kind: (definition as DefinitionMembers).Kind,
+	};
+
+	table.freeze(loggingDefinition);
+
+	return loggingDefinition;
+}
+
+export = createLoggingDefinition;