+++ src/Util/netBuilderFormat.ts
@@ -1,1 +1,4 @@
-export = (...params: unknown[]) => ["[netbuilder]", ...params] as LuaTuple<[string, ...unknown[]]>;
+import { DEFAULT_CONFIGURATION } from "./constants";
+
+export = (...params: unknown[]) =>
+	["[" + DEFAULT_CONFIGURATION.Label + "]", ...params] as LuaTuple<[string, ...unknown[]]>;